Rename Messaging class to TopicExchange
Rename the Messaging class to the more appropriate 'TopicExchange', and
make it solely concerned with instantiating a new object that represents
a RabbitMQ topic exchange.

Also make most of that class private; `TopicExchange#close` remains
public so that the calling class can close the connection when we no
longer need it.