Define defaults for CLI options
These are the same as the default AMQP host, username and password as
used by the Bunny library we're using to connect to RabbitMQ, but it's
best to be explicit.

Also, include the defaults in the option descriptions so that `optparse`
will display them when the command is called with `--help`.

This also means the `--topic` and `--exchange` options are now optional.