Rename TopicExchange class to AmqpClient
Rename the `TopicExchange` class to `AmqpClient` and move logic related
to topic exchanges into the `PublishUrls` class.

`TopicExchange` had too much responsibility for one class, it should be
concerned only with acting as an interface to RabbitMQ.

File rename to follow in a separate commit to keep the Git diff clean.