Reset the log level before each test
Because the logging level is set in `lib/govuk_seed_crawler.rb`, the
tests would randomly fail depending on the order in which they ran
(seeded randomised tests highlighted this). This was because the
GovukSeedCrawler module is not torn down between each test.

By resetting `Logger#level` to our default before each test, we avoid
that race condition.