Make sure correct parameters are passed through to Bunny for AMQP
This makes sure we use the correct connection parameters for our
`Bunny` object. The main issue was that our symbols were prepended
with `amqp_` which the library docs don't suggest.

As a result, when running this tool from the CLI it wouldn't honour
the parameters we pass in at the command-line and instead try to
connect using the default parameters that `Bunny` uses.