[#75177816] Allow setting the AMQP password using an ENV variable
We might not want to provide our passwords using the command line.
Instead, allow reading the password from an environment variable that
the configuration management tool controls.

In this case, the AMQP password will be read from
`GOVUK_CRAWLER_AMQP_PASS`.

Had to split the `#parse` method into two new methods: `#options` and
`#parse`. This is so that the original `Slop` object can be accessed in
our tests to peek inside of the `#help` method.