Add Jenkinsfile
This adds a basic Jenkinsfile to replace the previous scripts with the
new pipeline syntax. The pipeline is explicitly defined rather than
using the `buildProject` method as there are incompatibilities between
the project and that method (specifically problems due to incorrect gem
selection when using ruby 1.9.3 and different test set-ups).

It also sets the '.ruby-version' to 1.9.3-p550 which is the version that it
is run on in production.

As part of this work the 'publish-gem' task has been removed from rake
as this will now be handled by jenkins.