Add Rake version requirement
Rubygems/Bundler seems to struggle with this package, as I'm guessing
it doesn't take the Ruby version in to account when trying to
determine the latest version of rake which can be used.

To try and help it, lock the version to that which is "vendored" in
the 1.9.3 release of Ruby.