Require ffi 1.10.0
This is the last version of `ffi` which runs on Ruby 1.9.3 which is the
version we're running the seed crawler on in production. This prevents
an indirect dependency was upgrading to a more recent version of `ffi`,
which is a problem we're currently experiencing.

It looks like this was missed in
https://github.com/alphagov/govuk-puppet/commit/3485b4632046440d0d5f399d40fa3aa8b0757f3b
which meant that the version of ffi installed was being overwritten by
this Gem.