Bump to version 3.2.1
This is a bugfix that does not change the external interface of the tool, so I think a patch-level bump is appropriate.

Although we are bumping a dependency by a major version, this gem is intended to be used as an executable rather than as a library, and so I don't think this constitutes a breaking change.